Updated all ids to be long instead of int, updated version to 1.4.3

+++ api-client/src/main/java/com/desk/java/apiclient/model/BrandIds.java
@@ -34,11 +34,11 @@ public class BrandIds extends QueryFields {
 
     private BrandIds() {}
 
-    private BrandIds(Integer... fields) {
+    private BrandIds(Long... fields) {
         super((Object[])fields);
     }
 
-    public static BrandIds ids(Integer... ids) {
+    public static BrandIds ids(Long... ids) {
         return new BrandIds(ids);
     }
 }
