+++ rxjava-api-client/src/main/java/com/desk/java/apiclient/service/RxOpportunityService.java
@@ -75,7 +75,7 @@ public interface RxOpportunityService {
      * @return an opportunity api response observable
      */
     @GET(FILTERS_URI + "/{id}/" + OPPORTUNITIES_URI)
-    Observable<ApiResponse<Opportunity>> getOpportunitiesByFilterObservable(@Path("id") int filterId,
+    Observable<ApiResponse<Opportunity>> getOpportunitiesByFilterObservable(@Path("id") long filterId,
                                                                   @Query("per_page") int perPage,
                                                                   @Query("page") int page,
                                                                             @Query("embed") Embed embed);
@@ -88,7 +88,7 @@ public interface RxOpportunityService {
      * @return the opportunity observable
      */
     @GET(OPPORTUNITIES_URI + "/{id}")
-    Observable<Opportunity> getOpportunityObservable(@Path("id") int id, @Query("embed") Embed embed);
+    Observable<Opportunity> getOpportunityObservable(@Path("id") long id, @Query("embed") Embed embed);
 
     /**
      * Retrieve the opportunity timeline (history)
@@ -97,7 +97,7 @@ public interface RxOpportunityService {
      * @return the opportunity timeline observable
      */
     @GET(OPPORTUNITIES_URI + "/{id}/" + HISTORY_URI)
-    Observable<OpportunityTimeline> getOpportunityTimelineObservable(@Path("id") int opportunityId);
+    Observable<OpportunityTimeline> getOpportunityTimelineObservable(@Path("id") long opportunityId);
 
     /**
      * Retrieve opportunity activities and events.
@@ -108,7 +108,7 @@ public interface RxOpportunityService {
      * @return the activities and events api response observable
      */
     @GET(OPPORTUNITIES_URI + "/{id}/" + ACTIVITIES_URI)
-    Observable<ApiResponse<IOpportunityActivity>> getOpportunityActivitiesObservable(@Path("id") int opportunityId,
+    Observable<ApiResponse<IOpportunityActivity>> getOpportunityActivitiesObservable(@Path("id") long opportunityId,
                                                                                      @Query("per_page") int perPage,
                                                                                      @Query("page") int page);
 
@@ -138,5 +138,5 @@ public interface RxOpportunityService {
      * @return an attachment api response
      */
     @GET(OPPORTUNITIES_URI + "/{id}/" + ATTACHMENTS_URI)
-    Observable<ApiResponse<OpportunityAttachment>> getAttachmentsObservable(@Path("id") int opportunityId, @Query("per_page") int perPage, @Query("page") int page);
+    Observable<ApiResponse<OpportunityAttachment>> getAttachmentsObservable(@Path("id") long opportunityId, @Query("per_page") int perPage, @Query("page") int page);
 }
